update đã có thể thanh toán. chưa xử lý được phần bàn trả về sau khi thanh toán
