+++ src/main/java/com/codegym/entity/Desk.java
@@ -20,7 +20,6 @@ public class Desk {
     @GeneratedValue(strategy = GenerationType.AUTO)
     private Long id;
     private String name;
-
     private Boolean status;
 
     @JsonIgnore
