+++ src/main/java/com/codegym/repository/BillRepository.java
@@ -1,0 +1,7 @@
+package com.codegym.repository;
+
+import com.codegym.entity.Bill;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface BillRepository extends JpaRepository<Bill, Long> {
+}
