+++ src/main/java/com/codegym/repository/OrderRepository.java
@@ -1,5 +1,6 @@
 package com.codegym.repository;
 
+import com.codegym.entity.Desk;
 import com.codegym.entity.Order;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
@@ -10,4 +11,5 @@ import java.util.Optional;
 @Repository
 public interface OrderRepository extends JpaRepository<Order, Long> {
     Order getOrderByDesk_Id(Long id);
+
 }
