+++ src/main/java/com/codegym/service/bill/IBillService.java
@@ -1,0 +1,7 @@
+package com.codegym.service.bill;
+
+import com.codegym.entity.Bill;
+import com.codegym.service.IGeneralService;
+
+public interface IBillService extends IGeneralService<Bill> {
+}
