+++ src/main/webapp/WEB-INF/views/layout/desk.html
@@ -78,7 +78,7 @@
                         <input type="text" class="form-control" id="deskName" data-rule-required="true"
                                data-msg-required="Không được bỏ trống">
                     </div>
-                    <button type="button" class="btn btn-danger" onclick="createTable()">Tạo Bàn</button>
+                    <button type="button" class="btn btn-danger" onclick="createDesk()">Tạo Bàn</button>
                 </form>
             </div>
             <div class="col-xl-6">
@@ -108,6 +108,7 @@
                                     <h2 id="desk_name"></h2>
                                     <thead>
                                     <tr>
+                                        <th>#</th>
                                         <th>Tên món</th>
                                         <th>Giá</th>
                                         <th>Số lượng</th>
@@ -116,8 +117,8 @@
                                     </thead>
                                     <tbody id="product_list_of_desk">
                                     <!--Hiển thị danh mục món ăn của bàn-->
+                                    </tbody>
                                     <tfoot id="total">
-
                                     </tfoot>
                                 </table>
 
@@ -183,10 +184,51 @@
         </div>
     </div>
 </div>
+<!--Modal bill-->
+<!-- Modal -->
+<div class="modal fade" id="bill" tabindex="-1" role="dialog" aria-labelledby="exampleModalScrollableTitle" aria-hidden="true">
+    <div class="modal-dialog modal-dialog-scrollable" role="document">
+        <div class="modal-content">
+            <div class="modal-body">
+                <div class="row">
+                    <h1 style="text-align: center">NHÀ HÀNG 4 THÀNH VIÊN</h1>
+                    <h2 style="text-align: center" >ĐC: số 1 - Thành Phố Huế</h2>
+                    <h3 style="text-align: center">HOTLINE: 0909900990</h3>
+                    <p>-----------------------------------------------------------------------------------------------------</p>
+                    <form class="form-group">
+                        <input type="number" id="order_id" hidden>
+                        <label for="customerName">Tên khách</label>
+                        <input class="form-control" id="customerName" type="text">
+                        <label for="staffName">Tên NV</label>
+                        <input class="form-control" id="staffName" type="text">
+                        <table class="table table-hover">
+                            <h2 id="desk_name_bill"></h2>
+                            <thead>
+                            <tr>
+                                <th>#</th>
+                                <th>Tên món</th>
+                                <th>Giá</th>
+                                <th>Số lượng</th>
+                                <th>Thành tiền</th>
+                            </tr>
+                            </thead>
+                            <tbody id="product_list_of_bill">
+                            <!--Hiển thị danh mục món ăn của bàn-->
+                            </tbody>
+                            <tfoot id="total_bill">
+                            </tfoot>
+                        </table>
+                    </form>
+                </div>
+            </div>
+
+        </div>
+    </div>
+</div>
 
 <!--   Core JS Files-->
 
-<script src="/resource/js/desk/desk.js"></script>
+<script src="/resource/js/deskPage/desk.js"></script>
 <!--font-->
 <script defer src="https://use.fontawesome.com/releases/v5.15.4/js/all.js"></script>
 <script defer src="https://use.fontawesome.com/releases/v5.15.4/js/v4-shims.js"></script>
