+++ src/main/webapp/WEB-INF/views/user/admin.html
@@ -21,11 +21,11 @@
 </head>
 <body>
 <!--preloader-->
-<!--<div id='af-preloader'>-->
-<!--    <div class='af-preloader-wrap'>-->
-<!--        <div class='af-sp af-sp-wave'></div>-->
-<!--    </div>-->
-<!--</div>-->
+<div id='af-preloader'>
+    <div class='af-preloader-wrap'>
+        <div class='af-sp af-sp-wave'></div>
+    </div>
+</div>
 
     <!--    Start mainmenu area-->
     <section class="mainmenu-area stricky">
